// src/types.ts
export interface GroupMember {
  user_id: string;
  nickname: string;
  image_url: string | null;
  id: string;
  muted: boolean;
  autokicked: boolean;
  roles: string[];
  name: string;
}

export interface Group {
  id: string;
  group_id: string;
  name: string;
  phone_number: string;
  type: string;
  description: string;
  image_url: string | null;
  creator_user_id: string;
  created_at: number;
  updated_at: number;
  muted_until: number | null;
  office_mode: boolean;
  share_url: string | null;
  share_qr_code_url: string | null;
  members: GroupMember[];
  messages: {
    count: number;
    last_message_id: string;
    last_message_created_at: number;
    preview: {
      nickname: string;
      text: string;
      image_url: string | null;
      attachments: Attachment[];
    };
  };
  max_members: number;
}

export interface Attachment {
  type: 'image' | 'location' | 'split' | 'emoji' | 'mentions' | 'reply';
  url?: string;
  name?: string;
  lat?: string;
  lng?: string;
  token?: string;
  placeholder?: string;
  charmap?: number[][];
  user_ids?: string[];
  loci?: number[][];
  reply_id?: string;
}
// ...
export interface Chat {
  created_at: number;
  updated_at: number;
  messages_count: number;
  last_message: {
    attachments: Attachment[];
    avatar_url: string | null;
    conversation_id: string;
    created_at: number;
    favorited_by: string[];
    id: string;
    name: string;
    recipient_id: string;
    sender_id: string;
    sender_type: string;
    source_guid: string;
    text: string | null;
    user_id: string;
  };
  other_user: {
    avatar_url: string | null;
    id: string;
    name: string;
  };
}